A better Poc

+++ POC/DTO/EntireOrder.cs
@@ -4,7 +4,7 @@ public class EntireOrder
 {
     public string ConsumerName { get; set; }
     public DateTime OrderTime { get; set; }
-    public double TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
     public List<EachOrderItem> OrderItems { get; set; }
 }
 
@@ -13,5 +13,5 @@ public class EachOrderItem
     public string SupplierName { get; set; }
     public string ProductName { get; set; }
     public uint QuantityOrdered { get; set; }
-    public double UnitPrice { get; set; }
+    public decimal UnitPrice { get; set; }
 }
